feat: move invite collaborators step to main onboarding flow
Remove the invite collaborators step from the GHGI/inventory onboarding
and add it to the main city onboarding flow. Also fix city selection
state not resetting when all invited members are removed.
